refactory. 인풋 패키지 생성
