+++ src/main/java/racingcar/input/Attempt.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.input;
 
 import camp.nextstep.edu.missionutils.Console;
